Ingesta-MySQL/requirements.txt
mysql-connector-python
pandas
boto3
